refactor: move buttons to their own subfolder so that it's clear they're sub-components of the actions panel
