+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/ActionSettingsButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ActionSettingsButton } from "./ActionSettingsButton";
