+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/ChartSettingsButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ChartSettingsButton } from "./ChartSettingsButton";
