+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/LinkCardEditButton/index.ts
@@ -1,0 +1,1 @@
+export { LinkCardEditButton } from "./LinkCardEditButton";
